refactor: renamed to personal_graph
