--- libsql_graph_db/__init__.py
@@ -1,1 +1,0 @@
-from libsql_graph_db import graph
