+++ personal_graph/__init__.py
@@ -1,0 +1,3 @@
+from personal_graph import graph
+
+__all__ = ["graph"]
