+++ personal_graph/sql/delete-edge-embedding.sql
