+++ personal_graph/sql/delete-edge.sql
