+++ personal_graph/sql/delete-node-embedding.sql
