+++ personal_graph/sql/delete-node.sql
