+++ personal_graph/sql/insert-edge-embedding.sql
