+++ personal_graph/sql/insert-node-embedding.sql
