+++ personal_graph/sql/search-node.template
