+++ personal_graph/sql/vector-search-edge.sql
