+++ personal_graph/sql/vector-search-node.sql
