+++ scripts/time_complexity.py
@@ -10,8 +10,8 @@ script_dir = os.path.dirname(os.path.realpath(__file__))
 parent_dir = os.path.abspath(os.path.join(script_dir, os.pardir))
 sys.path.append(parent_dir)
 
-from libsql_graph_db import database as db  # noqa: E402
-from libsql_graph_db import natural as nt  # noqa: E402
+from personal_graph import database as db  # noqa: E402
+from personal_graph import natural as nt  # noqa: E402
 
 
 def measure_execution_time(operation_func, *args, **kwargs):
